Remove the SessionRequest struct by instead passing an owned flatbuffer message buffer
